FolderIconLoadTest Deflake
The FolderIconLoadTest has always been flakey because in the test when we are trying to wait for the icon cache to fully update before clearing cache, there was a period of time between when the last SerializedIconUpdateTask was updating and the next one was scheduled in which there was no task currently sitting in the queue. Instead of checking the queue for messages, we have changed this to use a volatile variable that is set when any update task is started, and set to false when all update tasks are no longer running.

I have also switched the objects from stacks to ArrayDequeue becase the use of Stacks is considered deprecated. Because ArrayDequeue pop() is a reverse order operation, I have also switched to using removeLast() to maintain parity with the previous implementation.

Flag: None
Test: FolderIconLoadTest
Bug: 319923578
Change-Id: Ie9b391a56d1651f867a132c4092b490cd8b3cd0e