Always Return AdaptiveIconDrawable
As part of the legacy icon format cleanup we are making normalizeAndWrapToAdaptiveIcon always return an adaptive icon so no further checks need to be made. We have also removed a optimization here for simplicity that would return the legacy icon if it was square and the icon shape was square.

Bug: 318867282
Flag: None
Test: AllAppsSearchImageTest
Change-Id: Id9bef4b99c618ed34d4d49257faa1e16807e65ed