Fix Weather Engine loading
(original cl: go/aux-gerrit/21340)

There were a several issues with the current implementation
of the WeatherEngine:
- We were not loading/saving the last weather. That made that
if the wallpaper is set and we boot the device, we don't have
a weather and thus we did not show anything.
- Assets loading were set in onResume. This caused a bunch of issues,
from not loading the assets to not having a preview of the wallpaper
(because the picker pauses and stops the wallpaper several time while
getting it ready to preview, thus the asset loading was
being cancelled).
- Not having a thumbnail was causing a crash on the picker.
- Moved the stopping of the loading job to onDestroy (for same
reason as stated before).
- First time debug activity was loaded it had missing wallpaper.
- Fixed issue with framebuffer.
- Intensity setting was not visible.

Flag: EXEMPT MP apk not in build yet
Bug: 352352241
Test: visual
Change-Id: I70770b26fdad69d48c86efe1adf0e0fbd946e954