Adding support for fallback haptic effects to MSDL Tokens.
In the event that the original haptic effect for a MSDL token is not
supported, the repository will provide a fallback that the player can
use to play less premium (but still consistent) haptics. For faster
operations, the knowledge about which tokens require a fallback has been
cached in the player as a Map.

Test: atest MSDLPlayerImplTest
Flag: NONE use of the player will be flagged but its corresponding use
  case
Bug: 344654090
Change-Id: I0c25741837a284f6052b0c6f3ae0063dc0b29138