Fix for Top Dogfood Backup/Restore issue.
Sometimes when deferred restore is occurring, the launcher is killed and
restarted in restricted mode. Backup Restore uses this mode to clear the
data from the nexuslauncher app. However, this can conflict with
launcher loading and requesting data from its DBs. The solution is to
use a try catch around this issue so the app doesn't crash, and then
rely upon later launcher processes to clean up the icons that weren't
deleted.

Bug: 353505773
Test: Reproduced this bug locally, then verified that the backup restore
happens successfully after the fix is introduced.
Flag: EXEMPT bug fix

Change-Id: I958bf78d9db26b678be44e83aafb40b41d2ff966